완료 후 메세지 출력 미완성

+++ src/main/java/com/sparta/post/service/PostService.java
@@ -17,6 +17,14 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
+    public List<PostResponseDto> getPosts() {
+        // DB 조회
+        return postRepository.findAllByOrderByModifiedAtDesc()
+                .stream()
+                .map(PostResponseDto::new)
+                .toList();
+    }
+
     public PostResponseDto createPost(PostRequestDto requestDto) {
         // RequestDto -> Entity
         Post post = new Post(requestDto);
@@ -30,34 +38,39 @@ public class PostService {
         return postResponseDto;
     }
 
-    public List<PostResponseDto> getPosts() {
-        // DB 조회
-        return postRepository.findAllByOrderByModifiedAtDesc().stream().map(PostResponseDto::new).toList();
-    }
-
+    @Transactional
     public Long updatePost(Long id, PostRequestDto requestDto) {
         // 해당 메모가 DB에 존재하는지 확인
         Post post = findPost(id);
 
-        // post 내용 수정
-        post.update(requestDto);
+        if (post.getPassword().equals(requestDto.getPassword())) {
+            // post 내용 수정
+            post.update(requestDto);
+        } else {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
 
         return id;
     }
 
     @Transactional
-    public Long deletePost(Long id) {
+    public Long deletePost(Long id, String password) {
         // 해당 메모가 DB에 존재하는지 확인
         Post post = findPost(id);
 
-        // post 삭제
-        postRepository.delete(post);
+        if (post.getPassword().equals(password)) {
+            // post 삭제
+            postRepository.delete(post);
+        } else {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
         return id;
     }
 
-    private Post findPost(Long id) {
+    public Post findPost(Long id) {
         return postRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("선택한 메모는 존재하지 않습니다.")
+                new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.")
         );
     }
 }
